improve network deletion, add delete action in table
