+++ @xen-orchestra/web-core/lib/tables/column-sets/network-columns.ts
@@ -1,7 +1,7 @@
 import { defineColumns } from '@core/packages/table/define-columns.ts'
+import { useActionColumn } from '@core/tables/column-definitions/action-column.ts'
 import { useLinkColumn } from '@core/tables/column-definitions/link-column'
 import { useNumberColumn } from '@core/tables/column-definitions/number-column.ts'
-import { useSelectItemColumn } from '@core/tables/column-definitions/select-item-column'
 import { useStatusColumn } from '@core/tables/column-definitions/status-column.ts'
 import { useTextColumn } from '@core/tables/column-definitions/text-column.ts'
 import { useTruncatedTextColumn } from '@core/tables/column-definitions/truncated-text-column'
@@ -17,6 +17,6 @@ export const useNetworkColumns = defineColumns(() => {
     vlan: useNumberColumn({ headerLabel: () => t('vlan') }),
     mtu: useNumberColumn({ headerLabel: () => t('mtu') }),
     defaultLockingMode: useTextColumn({ headerLabel: () => t('default-locking-mode') }),
-    selectItem: useSelectItemColumn(),
+    actions: useActionColumn({}),
   }
 })
